pyproject.toml
[project]
name = "taskhub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[all]>=2.11.4",
    "pyjwt>=2.10.1",
    "pylint>=3.3.7",
    "requests==2.32.3",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
